SDI-102:  Fixed imports
Fixed ordering of imports in collector_test so as to comply with
goimport.  The two tools (goimport and gofmt) result in different
formatting and if you don't introduce your own space between builtin,
remote and local imports gofmt will have a different result than
goimports.  To identify issues more easily in the future we will show a
diff when running test.sh and fail instead of writing the file.