Fixes, debugging features, and tests
Fix: Adding a namespace with a single item did not add the Node into the root and did not generate the keys byte array.
Change: Moved setting of keys to node.setkeys to make sure keys conversion to the keys byte was common.
Change: Added a Debug flag to ConfigTree and logging for debugging behavior.
Fix: The long namespace and short tree test was wrong. The root node should return in the response since it lies in the path of the long namespace.
Change: Added panic for calling Get() on a non frozen tree. Test added for this case.
Change: Exported Print() on ConfigTree
Fix: Fixed deadlock where we should have been using defer with mutex.Unlock() inside Add() @geauxvirtual pointed this out after we merged it.