Re: PR comments
When a mutex is embedded in a type, its receiver becomes that type.
I.e., when passing said type via value, it can be dangerous. However, if
all operations on the type are on a pointer to it, the reciever of
Lock() and Unlock() is a pointer to the type.