fixes #361
When the `task.spin()` goroutine returns, it leaves the one running
`task.waitForSchedule()` blocked.  When the task is restarted, its
response chan is available again, causing the hung `waitForSchedule`
goroutine to begin execution again.

At this time, the _new_ call to `spin()` spawns a _new_ goroutine for
`waitForSchedule`, causing the spin loop to be executed twice.

This pattern continued for each restart of a task.