adds https ability to go rest client and pulsectl
HTTPS is available in two modes on the REST API: with a given certificate
/ key pair, or it can generate unsigned ones at start up.  The client
 needs the ability to bypass cert signing errors when communicating
with the API.  In this commit a param is added when constructing a
client instance called `insecure`.  If this flag is given, the
certificating signing errors are bypassed.

When pulsectl constructs a client instance, it passes in the value of a
new `--insecure` flag which provides pulsectl the same functionality
described above.