start plugins synchronously
When a task starts it calls SubscribeDeps on control to prepare any
plugins it may need.  Previously, SubscribeDeps emitted events which
were picked up by the runner, who then started the plugins.  This
occasionally resulted in an error starting the task, because the async
process of starting the needed plugins had not completed yet.  This was
especially true with tasks with very small intervals, <250ms for
example.

This commit causes SubscribeDeps to first subscribe, then check that the
pool is eligible.  If it is, then it runs a plugin directly.  The runner
no longer handles `PluginSubscriptionEvent`.