fixes #804: give fixed names to gob registered types
When using vendoring in Go 1.6, reflection is in gob.Register to create
a key for type lookups.  This `reflect.PkgPath()` call returns a "globally
unique" key for looking up types, and in the case of a vendored type, is
not valid for matching gob registration names.  This is because on the
server side, the type is registered as `import/path.(typename)`, but on
the client side (vendored) it looks like
`client/vendor/import/path.(typename)`.  Using fixed keys bypasses the
need for reflection and is safe on either side, even when using
vendoring.