Allow remote targets for nodes in a task workflow
Allows remote targets for nodes in a task workflow via a 'Hostname'
field. When set the node that is set will run on the target host
specified by the field.

The clients for each node are generated on task creation. A client is
created for each node that specifies a Hostname.