Add http endpoints to use profiling tools
Add an optional flag "--pprof" for snap daemon. When this flag is set
the rest API add profiling tool routes. The flag will also be passed
when running a plugin. Now, response from the plugin have a field for
passin port used for profiling tools.

Made the change on snapctl to display ports on running plugins.