Add snap-plugin-collector-kubestate to plugin list
I'd like to add the [snap-plugin-collector-kubestate plugin](https://github.com/raintank/snap-plugin-collector-kubestate) to the plugin list. We have been using it in production for a more than a month and I think it is ready to be used by other people now. 

This collector collects metrics from Kubernetes that are not available via Heapster or the docker collector. It is based on the Prometheus collector https://github.com/kubernetes/kube-state-metrics