Initial attempt at a WiX installer
First attempt using WiX Editor. Generated simple installer that places the executable in Program Files (for now) and installs a service that starts automatically.

WiX Generated with VS

Snap themed installer that installs snap.exe and an application menu
shortcut to a selected directory. (We will probably want better images
in the long run)

MSI Updates

Added service for snapteld but it's returning errors because the path is
wrong or when it is correct it doesn't properly run in the background
and the installer times out thinking it's failed.